PYCBC-1320: acouchbase - add core API docs
Motivation
==========
We want the Python SDK API docs to be useful for our users.  The
acouchbase API docs were missing in the 4.0.0 releae due to time
constraints.  We are now making progress to completing the API docs for
the asynchronous APIs.

Modification
============
Docstring updates.

Results
=======
Modifications should not impact functionality.  So, all tests pass.

Change-Id: I087294a18ab1a9cbd8a4bc73a0534b0ed847bdf3
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/175720
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>