PYCBC-1374:  Add build flag to enable dynamic linking of stdc++ library
Motivation
==========
The static stdc++ libs are difficult to find on RHEL8, having a flag
that allows for dynamic linking enables the client to be built on RHEL8.
Adding a flag to allow for dynamic linking can help in other situations
as well.

Modification
============
Update build system to check for COUCHBASE_CXX_CLIENT_STATIC_STDLIB to
allow for dynamic linking of stdc++ libs.

Results
=======
Builds remain successful on other OSes and can build client on RHEL8.

Change-Id: Icbe9860a3f15528828a7171003e52ea9f77879fa
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/176450
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>