PYCBC-1369 and PYCBC-1368
Motivation
==========
Add pass-through logging and the missing parts of custom metadata
collections for transactions.   Had to go together as the updated
transaction lib had both changes.

Modification
============
Created custom sink to allow client and transaction logs to flow through
to the python logger.   Additionally, added the notion of a
TransactionKeyspace that encapsulates a bucket/scope/collection, which
we use to specify the custom metadata collection in both
TransactionOptions and TransactionConfig.   Note it can be initialized
from an existing collection, if one exists.

Added tests for the metadata collections in the TransactionOptions as
well.   No logging tests yet - worth doing soon.

Also updated client and transactions lib to latest.  Minor update to a
diagnostics test, as the cxx client now maps a 500 from a query service
a bit differently.

Results
=======
Passing tests, interface looks correct.   Adhoc testing of the logging
looks good.

Change-Id: Ifff61ea4cea56b11bb6e7ebcb04616f859c72b1f
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/176558
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>