PYCBC-1382: Custom Python types should call tp_free in their destructor
Motivation
==========
If tp_free is not called when a custom Python type is deallocated, the
object used for the type is not freed causing a memory leak.

Modification
============
Update the Couchbase Python SDK's custom types to call tp_free at the
end of the type's destructor.

Results
=======
No further memory leaks related to the SDK's custom types.

Change-Id: If2742147064e7104473d1e8ce420aa9bc42fdc3f
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/177635
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>