PYCBC-1390: Search metrics not forwarded to Search result's metadata
Motivation
==========
Users should have access to a Search query's metrics via the Search result's metadata. Although the metrics are handled from the cxx client, the are not properly forwarded so that end users can access the metrics.

Modification
============
Forward the metrics from a Search query result so that users have access
to the metrics via a method call on the Search result's metadata
object. Update tests to validate response metadata and metrics.

Results
=======
All tests pass.

Change-Id: I2d94e2821ba098bddbcfd3f753469764b28a6878
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/178239
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>