PYCBC-1261 External tracing API
Motivation
==========
Add ability to use an external tracer, such as OpenTelemetry.

Modification
============
We have a CouchbaseSpan and CouchbaseTracer which present an interface
that one could implement for some specific tracing API.   Additionally,
there is now an CouchbaseOtelSpan and CouchbaseOtelTracer to wrap the
Opentelemetry Span and Tracer, if one wants to use those in particular.

Added tests for KV, query, search, analytics and views.

Results
=======
Well, still working on why views are a problem, but the rest pass.  All
the leak checking I've done looks good as well

Change-Id: I2b7d485e5db2570dc025fb2dbca901605c33a50f
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/178573
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>