PYCBC-1398:  Fix crash when accessing base_exception error_context
Motivation
==========
When working with views, it is possible to experience a crash with the
SDK in certain scenarios when an exception can be raised. This behavior
should be fixed in order to prevent the SDK from crashing.

Modification
============
When building ViewErrorContext, fix incorrectly DECREF'd PyObject.
Added tests to verify streaming API (includes views) handles bad
queries/indexes correctly.

Results
=======
All tests pass.

Change-Id: I6c2baced1d2a4940ee78fbc21ba264fe29a248a4
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/178726
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>