PYCBC-1312: Intermitent segfaults when closing cluster
Motivation
==========
Though cluster.close() isn't _necessary_, since it closes in the
destructor, we still should not segfault.  Allows users to manage the
lifetimes of the underlying resources (like file descriptors).

Modification
============
The main issue was a missing PY_INCREF on the connection capsule before
passing it into the close callback.  This definitely causes issues with
calling close multiple times on the cluster as well.

Also, I make sure we close_connection in the destructor for
the ClusterLogic.   Testing locally, the connections are always
destroyed, whether or not close is called explicitly or not.

Added a test to call close multiple times as well, which verifies this
fixes PYCBC-1407 as well.

Results
=======
Tests pass.

Change-Id: Ie0d43241f0497f97ff799bb157a743bcd96e0dbd
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/179961
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>