PYCBC-1413: Streaming API can raise exception when executing in thread
Motivation
==========
The couchbase streaming API should work w/in applications that utilize
threading.  Previously the query API was corrected, but the fix was not
applied to the remainder of the streaming API (analytics, search,
views).  This ticket is to fix rest of the streaming API.

Modification
============
Move asyncio.Queue rows to acouchbase only.  Add thread tests to the
streaming API to confirm functionality.

Results
=======
All tests pass.

Change-Id: Iddd60f41d5d007e41d93e683e41580d147124905
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/180573
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>