PYCBC-1438: Update Python 4.x Test Framework
Motivation
==========
In order to provide more stability to the CI/CD pipeline the test
framework should be reworked to not rely on creating a connection for
each test suite.  The unneccesary number of connections can create
instability.

Modification
============
Reworked the test framework to have a global connection that can be
reused.  Only specific tests or test suites that need a new connection
create a connection now.  Creating of docs and removing has also been
updated so as to allow for bulk insert/removal further reducing the time
for tests to complete.

Results
=======
All tests pass.

Change-Id: I9f449b65bbcbe060699adfb7a0082e9f34ff3b13
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/186267
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>