PYCBC-1443:  Import ssl package prior to importing any couchbase package
Motivation
==========
In order to determine of OpenSSL is available, we attempt to import the
ssl package in couchbase/__init__.py first.  If that package can be
successfully imported, OpenSSL should be available and then we can
complete tasks that require imports from the couchbase package.

Modification
============
Move all imports from the couchbase package below the check for
importing the ssl package and make sure formatting and linting tools
will not move the imports to the top of the __init__.py file.

Results
=======
No import failures when working with the SDK (specifically on MacOS
machines).

Change-Id: Id618edfb6a30a79a3e36b723d2ceceb1880cb612
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/186656
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>