PYCBC-1469: Check if Python interpreter is finalizing prior to logging
Motivation
==========
It is possible, since we do not check, that the Python interpreter is in
the process of finalizing when we try to log out a message from spdlog.
If the interpreter is in the process of finalizing we could potentially
run into unwanted termination.  Also, when logging we have two instances
when the Py_None PyObject is used without incrementing the reference.
We should always increment Py_None when using.

Modification
============
Check if the Python interpreter is finalizing prior to attempting to add
a log message from the C++ client. Increment Py_None PyObject when
using.

Results
=======
Logging continues to work.  All tests pass.

Change-Id: Ibf101cf41b0fac44dca1328b5ea632ba211548e3
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/189367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>