PYCBC-1474: Transactions not handling DocumentNotFound within lambda correctly
Motivation
==========
Within a transactions lambda we should raise certain SDK3 exceptions
(DocumentNotFound, DocumentExists, etc.). However, we are currently
using a legacy Get stage w/in transaction operations that will return an
empty repsonse when trying to get a document that does not exist.  We
need TXNCXX-215 to be completed in the C++ client before we can remove
the use of the legacy Get stage.  So, in the interim we need to check
for a non-error and empty response and raise a DocumentNotFoundException
accordingly.

Modification
============
Update the bindings to properly handle various error situations so that
we raise the correct error within the transactions lambda.  Added tests
in order to verify updated functionality and that we raise the correct
errors at various stages within transactions operations.

Results
=======
All txn tests pass.

Change-Id: I5e550d9a086a396f5fc6ba6e906c884025419739
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/190116
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>