PYCBC-1475: Allow exceptions to be created with only a positional message argument
Motivation
==========
Often times an exception is raised w/ just a message as a positional
argument.  We should allow that behavior for all our exceptions and make
sure that when a user gets the str or repr of the exception we output a
useful string representation of the exception.

Modification
============
Update exception classes to to allow for an optional positional
argument, update exception classes to have appropriate __str__ and
__repr__ methods.  Add test to verify.

Results
=======
All tests pass.

Change-Id: I23e83551a75f0f0b456841c0c7142b2b5c0eed76
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/190137
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>