PYCBC-1477: Fix possible MutationState crash
Motivation
==========
Whe doing read your own write query/search operations, mutation state
needs to be passed in as an option to the query/search.  Due to how the
mutation_tokens were passed in as an option to the binds it was possible
to run into a potential crash.

Modification
============
Convert mutation tokens into their dict form which is what is
expected in the bindings as we parse the mutation token into the
appropriate request for the c++ client. Update tests to test against
mutation state creation that would potentially allow a crash to confirm
we don't crash.

Results
=======
All tests pass.

Change-Id: I3ddfb963ec3e33f09093a2ad9e97bc1af5873b79
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/190193
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>