PYCBC-1471:  acouchbase streaming APIs block while executing queries
Motivation
==========
The acouchbase API should be a non-blocking API for most (if not all) operations.
Currently the streaming APIs (analytics, query, search and views) will
block while retrieving results from a query.  This behavior should be
corrected so that the streaming APIs are non-blocking.

Modification
============
Since the underlying operating to retrieve a row from the bindings is a
blocking operation (pulling an object from a results queue), update to
running the operation in another thread via a thread pool executor. Add
appropriate tests to verify behavior where possible (unfortunately the
behavior is hard to see w/o queries that are slow).

Results
=======
The acouchbase streaming APIs are no longer blocking and all tests pass.

Change-Id: Iea7e5712aa60598ea0cd3cdd38030ec4bc8a94f8
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/190253
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>