PYCBC-1480: Subdocument lookup-in should correctly handle null values.
Motivation
==========
The Python SDK's lookup-in operation is not handling documents that have
null values for a specified path.  We should update the lookup-in logic to handle
null values correctly (i.e. properly pass them along as a null value is valid
JSON) and to correctly follow the RFC.

Modification
============
Update LookupInResult content_as() and exists() methods to correctly follow RFC.
Add tests to confirm updated logic, specifically that null values in a
document are handled correctly.

Results
=======
All tests pass.

Change-Id: Iebf2b9caff804a58cf74984bc344df6caaa65922
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/192048
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>