PYCBC-1511:  Provide access to protocol logger
Motivation
==========
The C++ core 1.0.0-dp.7 added support for a protocol logger (enables
network traffic to be saved to a file).  Exposing the protocol logger
would be useful for debugging in certain scenarios. NOTE:  this is a
volatile feature that is subject to future change/improvements.

Modification
============
Expose C++ core protocol logger.

Results
=======
Access to the C++ core protocol logger now available.

Change-Id: Ia9e8a0804491808b15a8cdced514ade4be0ca2eb
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/194699
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>