PYCBC-1521: Streaming APIs do not use cluster timeout values
Motivation
==========
Currently the Python SDK sets the timeout value for a streaming
operation on every request.  If a timeout value was not provided the
default 75s value was used instead of the cluster timeout value (if it
was specified).  The Python SDK should honor the cluster level timeouts
if specified.

Modification
============
Update the SDK to use the cluster level timeouts for streaming
operations.  Added tests to confirm functionality.

Results
=======
All tests pass.

Change-Id: I816d4e61f234c39ad60960617b7dd28ba8204b04
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/197151
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>