PYCBC-1538: Get should not fail with InvalidArgumentException when projecting on more than 16 fields
Motivation
==========
The RFC states that once the 16 field limit is reached the SDK should fall back to a full document fetch. Currently the Python SDK is raising an InvalidArgumentException. The C++ core already has the correct logic, so only the check for the number of fields in the projection needs to be removed

Modifications
=============
Remove the check for the number of fields in the projection
Update test_project_too_many_projections for all APIs

Results
=======
All tests pass

Change-Id: Ie9ece154d7269908f0b47520fc4d0fda34b76693
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/200237
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>