When getting the error context handle the case where the exception does not come from pycbc_core
Motivation
==========
If the error context is not set and the exception does not come from pycbc_core, calling an exception's error_context() method results in a "AttributeError: 'NoneType' object has no attribute 'error_context'".

Changes
=======
In the error_context getter, if the _base attribute of a CouchbaseException is None and _context is None, return an empty error context.

Change-Id: I0b37a81d4369dbd920665b95be1eb59d34f20e61
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/203433
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>