PYCBC-1560: Update ViewOptions to include full_set and raw
Motivation
==========
As more users migrate from older SDKs we should make sure the options
offered in the 4.x version of the SDK match what was offered in previous
versions (if the option is not deprecated).

Modification
============
Update ViewOptions to include/allow full_set and raw options.  Update
bindings to pass full_set and/or raw options to C++ core.  Add and
update tests to confirm functionality.

Results
=======
All tests pass.

Change-Id: Ibf9524a11f11bac3480e3c111428d5d36884ea47
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/205911
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>