PYCBC-1590: Cleanup logging on shutdown
Motivation
==========
When not using the console logger, there is still a possibility that we
try to handle some log messages from the C++ logger after the Python
interpreter has started to finalize.  This can cause the SDK to crash.
It is at the end of a script, but still unpleasant user experience.

Modification
============
Use Python's atexit() hook to call shutdown on the C++ core logger if
using logging via Python logging (console logger not impacted).

Results
=======
Console logger still provides output as expect and Python logger no
longer crashes on exit (worst-case is we rethrow the exception and that
is what we see in stderr/stdout).

Change-Id: I03fb4a4ce842c2e47f6396d05faaeb3f825bb720
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/208341
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>