PYCBC-1583: Fix BucketType.COUCHBASE being ignored from bucket settings
Motivation
==========
If bucket_type in BucketSettings is set to BucketType.COUCHBASE it is being ignored when converting to the C++ bucket settings. The reason is that the C++ wrapper code checks for the "couchbase" value instead of "membase". While this should not affect the overall behaviour, since COUCHBASE is also the server default, we should still send it if it's specified.

Change
======
Map both "membase" and "couchbase" to couchbase::core::management::cluster::bucket_type::couchbase in the C++ wrapper code.

Results
=======
createBucketWithAllSettingsAndVerifyAllSettings in FIT passes

Change-Id: I55c7511f57c36dacf811098110da1f63cb66ded9
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/210567
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>