PYCBC-1569: Provide way in increment/decrement to not create document if it does not exist
Motivation
==========
Currently, it is not possible to have an unset value for 'initial' in increment/decrement in the request sent to the server, as the initial value is initialized to 0 if not set in the options. The C++ core supports this, so we should provide an input that maps to unset initial in the core request.

Change
======
If the value of initial in increment/decrement is negative, map this to 'unset' in the core request. This means that DocuemntNotFoundException will be raised if the document does not exist.

Results
=======
All tests pass

Change-Id: I0e2b0889707e3020b2b2a5b0811acf7347abe098
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/217503
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>