Do not install cmake from PyPI when system CMake is available
Replace the unconditional build requirement on `cmake` with one that
is added only if no system executable can be found.  This ensures
that the system CMake version is used when available, and this can
be important when system CMake carries downstream patches for the
platform.  When no CMake executable is found, `cmake` is pulled in
via `setup_requires`, and the previous behavior is preserved.

Note that in order to accommodate this change, I had to remove the early
check for CMAKE_EXE being present.  This means that it is now possible
to use `setup.py` without having CMake installed, and it will only fail
when actually trying to compile the extension.

Change-Id: I3bc31206285168952a7b3f98d2c43a3997f95801
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/222207
Reviewed-by: Jared Casey <jared.casey@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>