PYCBC-1685: Remove timeout logic when waiting for C++ core HTTP response
Motivation
==========
Since the SDK uses a timeout (the same it passes down to the C++ core)
when waiting for streaming results from the C++ core, it is possible for
it to raise a `StopIteration` incorrectly.  Instead the SDK should
continue to wait for the C++ core to return a response.

Modification
============
* Remove logic to raise a `StopIteration` exception when waiting for a
  response for streaming operations.  Instead a message is logged (at
  the DEBUG level) to provide potentially useful information.
* Update tests to confirm changes.

Change-Id: Ia4ea33ebe9dbf3560874a5f9317b476a5ec5315a
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/227431
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>