PYCBC-1729: Skip tracing and metrics tests until new observability changes are complete
Motivation
==========
The underlying C++ core changes related to observability improvements
(CBD-6492) are blocking CI runs from passing as the old tracing/metrics
tests are failing.  The improvement work for updating how observability
is handled in the Python client is underway so these tests will
eventually be improved/updated.  Skipping the tests in the interim will
allow other work on the Python client to continue.

Modification
============
* Skip all tests w/in ClassicMetricsTests and ClassTracingTests

Results
=======
* CI runs no longer blocked by failing tracing/metrics tests

Change-Id: Ie8343f2999d35931d0ed3f5f11e3333faa368be6
Reviewed-on: https://review.couchbase.org/c/couchbase-python-client/+/236877
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>